requirements.txt
python>=3.14
pymupdf>=1.25.0
python-docx>=1.1.0
faiss-cpu>=1.9.0
neo4j>=5.0.0
langchain>=0.3.0
langchain-community>=0.3.0
sentence-transformers>=2.7.0
numpy>=1.26.0
tiktoken>=0.7.0
pytest>=9.0.0
rank-bm25>=0.2.0
pyyaml>=6.0
fastapi>=0.100.0
uvicorn>=0.25.0
